This is the second project - Investigate a Dataset (TMDb Movie Data)
For data analysis, I picked up and selected the TMDb movie dataset. This dataset is required to carry out all of the Data Anaysis's technique and its methodology and it consists of information from The Movie Database (TMDb) regarding 10,000 movies, including user reviews and revenue. Also, there is often too much information available to make a clear decision. that is why I have chosen to search for more interesting trends in the results.